requirements.txt
# Versiones compatibles para Windows
# Instala en orden: primero torch, luego ultralytics, después opencv

# PyTorch (CPU) - instalar primero
torch>=1.13.0
torchvision>=0.14.0

# Procesamiento de imágenes
opencv-python>=4.6.0
Pillow>=9.0.0

# Machine Learning
ultralytics>=8.0.0

# Utilidades científicas
numpy>=1.21.0
scipy>=1.9.0

# Visualización (opcional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Datos
pandas>=1.5.0

# Configuración
pyyaml>=6.0
tqdm>=4.60.0 